+++ lib/honeybadger/init/sinatra.rb
@@ -4,12 +4,7 @@ module Honeybadger
       ::Sinatra::Base.class_eval do
         class << self
           def build_with_honeybadger(*args, &block)
-            config = Honeybadger::Config.new(honeybadger_config(self))
-            if Honeybadger.start(config)
-              use(Honeybadger::Rack::ErrorNotifier, config) if config.feature?(:notices) && config[:'exceptions.enabled']
-              use(Honeybadger::Rack::MetricsReporter, config) if config.feature?(:metrics) && config[:'metrics.enabled']
-            end
-
+            install_honeybadger
             build_without_honeybadger(*args, &block)
           end
           alias :build_without_honeybadger :build
@@ -19,6 +14,21 @@ module Honeybadger
             {
               api_key: defined?(honeybadger_api_key) ? honeybadger_api_key : nil
             }
+          end
+
+          def install_honeybadger
+            config = Honeybadger::Config.new(honeybadger_config(self))
+
+            return unless config[:'sinatra.enabled']
+            return unless Honeybadger.start(config)
+
+            install_honeybadger_middleware(Honeybadger::Rack::ErrorNotifier, config) if config.feature?(:notices) && config[:'exceptions.enabled']
+            install_honeybadger_middleware(Honeybadger::Rack::MetricsReporter, config) if config.feature?(:metrics) && config[:'metrics.enabled']
+          end
+
+          def install_honeybadger_middleware(klass, config)
+            return if middleware.any? {|m| m[0] == klass }
+            use(klass, config)
           end
         end
       end